fixing index on relations
