+++ packages/twenty-server/src/engine/workspace-manager/workspace-sync-metadata/factories/standard-index.factory.ts
@@ -102,7 +102,9 @@ export class StandardIndexFactory {
           const hasAllFields = workspaceIndexMetadataArgs.columns.every(
             (expectedField) => {
               return objectMetadata.fields.some(
-                (field) => field.name === expectedField,
+                (field) =>
+                  field.name === expectedField ||
+                  `${field.name}Id` === expectedField,
               );
             },
           );
